admin/requirements.txt
python-dotenv==1.0.1
streamlit==1.45.1
pymongo==4.12.1
pandas==2.2.3
httpx==0.28.1
streamlit_oauth==0.1.14
python-magic==0.4.27
tiktoken==0.9.0
pillow==11.2.1
